refactor: manual link test isolated
